Add Vagrantfile for Handy Provisioning
This Vagrantfile spins up a Fedora 40 VM and performs the various steps
described in the Mad Scientist's Guide to Automating CNI Configurations
using Generative AI. For the sake of convenience, Ollama is run locally,
but the scripts can be adjusted to use an external host. It has been
tested only with the libvirt Vagrant provider but should be easily
portable to any provider supported by Vagrant.

vagrant up && vagrant ssh
~/start_cluster.sh
~/run_robocni.sh -h

Signed-off-by: Flavio Fernandes <flavio@flaviof.com>